cleaning gems and cats
