+++ packages/core/test/multiGiveaway/v2/fixtures.ts
@@ -36,7 +36,7 @@ type Options = {
 };
 
 export const setupTestGiveaway = withSnapshot(
-  ['PolygonAssetERC1155', 'PolygonGems', 'PolygonSand', 'PolygonCatalysts'],
+  ['PolygonAssetERC1155', 'PolygonSand'],
   async function (hre, options?: Options) {
     const {
       mint,
@@ -52,18 +52,12 @@ export const setupTestGiveaway = withSnapshot(
       assetBouncerAdmin,
       landAdmin,
       sandAdmin,
-      catalystMinter,
-      gemMinter,
       multiGiveawayAdmin,
     } = await getNamedAccounts();
     const otherAccounts = await getUnnamedAccounts();
     const others = otherAccounts;
     const sandContract = await ethers.getContract('PolygonSand');
     const assetContract = await ethers.getContract('PolygonAssetERC1155');
-    const speedGemContract = await ethers.getContract('PolygonGem_SPEED');
-    const rareCatalystContract = await ethers.getContract(
-      'PolygonCatalyst_RARE'
-    );
 
     await deployments.deploy('TestMetaTxForwarder', {
       from: deployer,
@@ -117,14 +111,6 @@ export const setupTestGiveaway = withSnapshot(
     if (sand) {
       await sandContractAsAdmin.transfer(giveawayContract.address, SAND_AMOUNT);
     }
-
-    // Supply Catalysts and Gems
-    await speedGemContract
-      .connect(ethers.provider.getSigner(gemMinter))
-      .mint(giveawayContract.address, 16);
-    await rareCatalystContract
-      .connect(ethers.provider.getSigner(catalystMinter))
-      .mint(giveawayContract.address, 8);
 
     // Supply assets
     const assetContractAsBouncerAdmin = await ethers.getContract(
@@ -235,8 +221,6 @@ export const setupTestGiveaway = withSnapshot(
           if (claim.erc20.amounts.length === 3)
             claim.erc20.contractAddresses = [
               sandContract.address,
-              speedGemContract.address,
-              rareCatalystContract.address,
             ];
         }
         return claim;
@@ -429,8 +413,6 @@ export const setupTestGiveaway = withSnapshot(
       sandContract,
       assetContract,
       landContract,
-      speedGemContract,
-      rareCatalystContract,
       others,
       allTrees,
       allClaims,
